# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/pregnancy-lactation": {
      "get": {
        "description": "2015 FDA PLLR structured-narrative format + LactMed-style lactation detail. Each record is keyed by drug slug and carries a clinical headline + risk summary + clinical considerations + data + reproductive-potential guidance. The retired A/B/C/D/X letter categories are stored for back-compat but should NOT be the primary decision aid. Filters: q (free text), pregnancy / lactation (preferred|acceptable|caution|avoid|contraindicated).",
        "operationId": "PregnancyLactationController_list",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pregnancy",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lactation",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List pregnancy + lactation reference records (PLLR format)",
        "tags": [
          "pregnancy-lactation"
        ]
      }
    },
    "/api/v1/pregnancy-lactation/{slug}": {
      "get": {
        "description": "Returns the full PLLR narrative: pregnancy (risk summary, clinical considerations, data, trimester notes), lactation (risk summary, infant exposure data, alternatives), reproductive potential (pregnancy testing, contraception, infertility), and citations.",
        "operationId": "PregnancyLactationController_get",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a pregnancy + lactation record by slug",
        "tags": [
          "pregnancy-lactation"
        ]
      }
    },
    "/api/v1/pregnancy-lactation/by-drug/{drugSlug}": {
      "get": {
        "description": "For an integrator that has the drug slug (e.g. from a prescription) and wants the PLLR record directly. Returns 404 if no PLLR record exists for that drug.",
        "operationId": "PregnancyLactationController_getByDrug",
        "parameters": [
          {
            "name": "drugSlug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reverse lookup by drug slug",
        "tags": [
          "pregnancy-lactation"
        ]
      }
    },
...
    "/api/v1/reference/summary": {
      "get": {
        "operationId": "ReferenceController_summary",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reference home — content counts + chat availability",
        "tags": [
          "reference"
        ]
      }
    },
    "/api/v1/reference/index": {
      "get": {
        "operationId": "ReferenceController_index",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "A–Z index across drugs, conditions + procedures (optionally filter by kind)",
        "tags": [
          "reference"
        ]
      }
    },
    "/api/v1/reference/groups/{kind}/{dimension}": {
      "get": {
        "operationId": "ReferenceController_groups",
        "parameters": [
          {
            "name": "dimension",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Grouped browse. drug: class|aware|keml. condition/procedure: specialty|domain (by clinical domain tags).",
        "tags": [
          "reference"
        ]
      }
    },
    "/api/v1/reference/search": {
      "get": {
        "operationId": "ReferenceController_search",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Free-text reference search across titles, slugs, classes + tags",
        "tags": [
          "reference"
        ]
      }
    },
    "/api/v1/reference/diagram/condition/{slug}": {
      "get": {
        "operationId": "ReferenceController_conditionDiagram",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mermaid management flowchart for a condition (deterministic, from signed content)",
        "tags": [
          "reference"
        ]
      }
    },
    "/api/v1/reference/diagram/procedure/{slug}": {
      "get": {
        "operationId": "ReferenceController_procedureDiagram",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mermaid step flowchart for a procedure (safety-check gates highlighted)",
        "tags": [
          "reference"
        ]
      }
    },
    "/api/v1/reference/diagram/interactions": {
      "post": {
        "operationId": "ReferenceController_interactionNetwork",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InteractionNetworkDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Drug-interaction NETWORK for a medication list (polypharmacy review) — Mermaid graph with severity-coded edges + a structured interaction list.",
        "tags": [
          "reference"
        ]
      }
    },
    "/api/v1/reference/diagram/drug-interactions/{slug}": {
      "get": {
        "operationId": "ReferenceController_drugInteractionMap",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Single-drug interaction map — the drug + all its interaction partners, severity-coded",
        "tags": [
          "reference"
        ]
      }
    },
    "/api/v1/reference/dose": {
      "post": {
        "operationId": "ReferenceController_doseCalc",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoseCalcDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Dose calculator — weight/age/renal-based, computed from the signed bundle. Refuses + cites when inputs missing or renally contraindicated.",
        "tags": [
          "reference"
        ]
      }
    },
    "/api/v1/reference/chat": {
      "post": {
        "operationId": "ReferenceController_referenceChat",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReferenceChatDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Guarded clinician reference chat — answers ONLY from VedaMD references, cites every claim, refuses anything not covered or non-medical. No patient data; PHI-free.",
        "tags": [
          "reference"
        ]
      }
    },
    "/api/v1/usage": {
      "get": {
        "operationId": "UsageController_list",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "since",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Recent usage events (newest first) for the calling integrator",
        "tags": [
          "usage"
        ]
      }
    },
    "/api/v1/usage/summary": {
      "get": {
        "operationId": "UsageController_summary",
        "parameters": [
          {
            "name": "since",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Usage rollup (by category, endpoint, day, actor type) for the dashboard",
        "tags": [
          "usage"
        ]
      }
    },
    "/api/v1/usage/events/client": {
      "post": {
        "description": "Whitelist-gated: only the fixed enum of `event` names is accepted; metadata is NOT stored. PHI-free.",
        "operationId": "UsageController_recordClientEvent",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Record a client-side feature use (calculator run, page view).",
        "tags": [
          "usage"
        ]
      }
    },
...
  },
  "info": {
    "title": "VedaMD API",
    "description": "Source-available, stateless Clinical Decision Support API. CDS Hooks + REST/JSON. VedaMD never stores patient data — records remain with your EMR.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "InteractionNetworkDto": {
        "type": "object",
        "properties": {}
      },
      "DoseCalcDto": {
        "type": "object",
        "properties": {}
      },
      "ReferenceChatDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}